create ArtworkAddInterface component
