+++ src/components/ArtworkAddInterface/_ArtworkAddInterface.scss
@@ -1,0 +1,28 @@
+@use "/src/utilities/variables";
+@use "/src/utilities/mixins";
+
+.add-interface {
+  @include mixins.formContainer();
+  position: relative;
+  width: 100%;
+  background-color: variables.$container-gray;
+
+  &__label {
+    display: none;
+  }
+
+  &__select {
+    @include mixins.searchSelect();
+    width: 200px;
+    height: 35px;
+    margin-right: 20px;
+  }
+
+  &__button {
+    @include mixins.themeButton();
+    font-size: .9rem;
+    height: 25px;
+    width: 87px;
+    margin-right: 15px;
+  }
+}
