+++ src/components/ArtworkDetails/ArtworkDetails.js
@@ -8,9 +8,9 @@ import { cleanDetails } from "../../utilities/cleaners"
 import { getArtworkDetails } from "../../utilities/apiCalls"
 
 import "./_ArtworkDetails.scss"
-import fallbackIMG from "../../assets/fallback.png"
 import SuccessAnimation from "../SuccessAnimation/SuccessAnimation"
 import ArtworkDetailsGallery from "../ArtworkDetailsGallery/ArtworkDetailsGallery"
+import ArtworkAddInterface from "../ArtworkAddInterface/ArtworkAddInterface"
 
 function ArtworkDetail() {
   const dispatch = useDispatch()
@@ -21,8 +21,6 @@ function ArtworkDetail() {
   const [error, setError] = useState(null)
   const [showSuccess, setShowSuccess] = useState(false)
   const [isLoading, setIsLoading] = useState(true)
-  const [selectedCollection, setSelectedCollection] = useState("")
-  // const [currentImg, setCurrentImg] = useState("")
   const [artworkData, setArtworkData] = useState({})
 
   const { additionalImages, artistName, artistURL, classification, country, culture, creditLine, department, description, geographyType, primaryImage, imageSmall, medium, objectDate, objectName, metURL, period, region } = artworkData
@@ -43,11 +41,23 @@ function ArtworkDetail() {
       if (!response.ok) throw Error(response.statusText)
       const details = await response.json()
       setArtworkData(cleanDetails(details))
-      // setCurrentImg(details.primaryImage ? details.primaryImage : fallbackIMG)
       setIsLoading(false)
     } catch (error) {
       setError(error)
     }
+  }
+
+  const handleDispatch = id => {
+    dispatch(addToCollection({
+      collectionID: id,
+      artistName,
+      culture,
+      department,
+      objectDate,
+      objectID,
+      objectName,
+      imageSmall
+    }))
   }
 
   const relatedCollections = useMemo(() => collections.reduce((acc, collection) => {
@@ -62,50 +72,10 @@ function ArtworkDetail() {
   }, []),
     [collections])
 
-  // const togglePhoto = newURL => {
-    // const targetIndex = additionalImages.indexOf(newURL)
-    // additionalImages.splice(targetIndex, 1, currentImg)
-  //   setCurrentImg(newURL)
-  // }
-
-  const handleSubmit = () => {
-    if (!selectedCollection) return
-    const targetCollection = collections.find(collection => collection.name === selectedCollection)
-    dispatch(addToCollection({
-      collectionID: targetCollection.id,
-      artistName,
-      culture,
-      department,
-      objectDate,
-      objectID,
-      objectName,
-      imageSmall
-    }))
-    setSelectedCollection("")
-    removeOption(targetCollection.id)
-    alertSuccess()
-  }
-
   const alertSuccess = () => {
     setShowSuccess(true)
     setTimeout(setShowSuccess, 3000, false)
   }
-
-  const removeOption = targetCollectionID => {
-    const targetIndex = relatedCollections.findIndex(collection => collection.id == targetCollectionID)
-    relatedCollections.splice(targetIndex, 1)
-  }
-
-  const collectionOptions = collections.filter(collection => {
-    return relatedCollections.every(related => related.id !== collection.id)
-  })
-    .map(collection => (
-      <option
-        value={collection.name}
-        key={collection.id}
-        data-cy={`option-${collection.id}`}
-      >{collection.name}</option>
-    ))
 
   const prevSavedString = relatedCollections.length > 1 ?
     `"${relatedCollections[0]?.name}" & others` :
@@ -132,27 +102,6 @@ function ArtworkDetail() {
 
   const artistSearchPath = `/search?query=${artistName?.replace(/ /g, "+")}&type=artist&dept=all&page=1`
 
-  // const addlPhotoButtons = !isLoading &&
-  //   additionalImages.reduce((acc, url, i) => {
-  //     if (i < 5) {
-  //       acc.push(
-  //         <button
-  //           className="artwork__right__img-controls__button"
-  //           onClick={() => togglePhoto(url)}
-  //           key={i + 1}
-  //           data-cy="image-button"
-  //         >
-  //           <img
-  //             className="artwork__right__img-controls__button__img"
-  //             src={url}
-  //             data-cy={`image-tile-${i + 1}`}
-  //           />
-  //         </button>
-  //       )
-  //     }
-  //     return acc
-  //   }, [])
-
   return (
     <div className="artwork-view-parent">
       <span className="artwork-directory">
@@ -176,7 +125,7 @@ function ArtworkDetail() {
             className="artwork__left__title"
             data-cy="object-title"
           >{objectName}
-          <SuccessAnimation showSuccess={showSuccess} />
+            <SuccessAnimation showSuccess={showSuccess} />
           </h3>
           <p
             className="artwork__left__date"
@@ -248,43 +197,16 @@ function ArtworkDetail() {
               data-cy="object-met-link"
             >metmuseum.org</a>
           </p>
-          <div className="artwork__left__collection">
-            <label
-              className="artwork__left__collection__label"
-              htmlFor="selected-collection"
-            >add to collection:</label>
-            <select
-              className="artwork__left__collection__select"
-              id="selected-collection"
-              value={selectedCollection}
-              onChange={e => setSelectedCollection(e.target.value)}
-              required={true}
-              data-cy="add-collection-select"
-            >
-              <option value="add to collection">add to collection</option>
-              {collectionOptions}
-            </select>
-            <button
-              className="artwork__left__collection__button"
-              onClick={handleSubmit}
-              data-cy="add-collection-submit"
-            >add</button>
-          </div>
+          <ArtworkAddInterface
+            alertSuccess={alertSuccess}
+            handleDispatch={handleDispatch}
+            collections={collections}
+            relatedCollections={relatedCollections}
+          />
         </div>
         <div className="artwork__right">
-          {/* <div className="artwork__right__mat">
-            {
-              !!currentImg &&
-              <img
-                className="artwork__right__mat__img"
-                src={currentImg}
-                alt={`${objectName}${artistName && ` by ${artistName}`}`}
-                data-cy="object-image"
-              />
-            }
-          </div> */}
-          { 
-            !isLoading && 
+          {
+            !isLoading &&
             <ArtworkDetailsGallery
               primaryImage={primaryImage}
               additionalImages={additionalImages}
@@ -292,9 +214,6 @@ function ArtworkDetail() {
               artistName={artistName}
             />
           }
-          {/* <div className="artwork__right__img-controls">
-            {addlPhotoButtons}
-          </div> */}
         </div>
       </section>
     </div>
