+++ src/components/ArtworkDetails/_ArtworkDetails.scss
@@ -84,32 +84,6 @@
   &__met-link {
     margin-bottom: 35px;
   }
-
-  &__collection {
-    @include mixins.formContainer();
-    position: relative;
-    width: 100%;
-    background-color: variables.$container-gray;
-
-    &__label {
-      display: none;
-    }
-  
-    &__select {
-      @include mixins.searchSelect();
-      width: 200px;
-      height: 35px;
-      margin-right: 20px;
-    }
-
-    &__button {
-      @include mixins.themeButton();
-      font-size: .9rem;
-      height: 25px;
-      width: 87px;
-      margin-right: 15px;
-    }
-  }
 }
 
 .artwork__right {
